+++ NEWS.md
@@ -5,10 +5,10 @@
 * Allow to point to .py configuration instead of project on command line.
   This means you can potentially have a farm of .py site descriptions pointing
   at various other directories in the file system.
-* archetypes directory configurable in `settings.py`. See
+* archetypes and output directory configurable in `settings.py`. See
   [settings.md](doc/settings.md) for details.
-* Added `--theme`, `--content` and `--archetypes` to command line to override
-  the corresponding settings.
+* Added `--theme`, `--content`, `--archetypes` and `--output` to command line
+  to override the corresponding settings.
 * Fixed a bug in taxonomy generation
 
 # New in version 0.2